chore(app,api): upgrade dependencies for better compatibility and performance
Updated various dependencies across the `api` and `app` packages to their latest versions to ensure better compatibility and performance enhancements. This includes critical updates to AWS SDK, Storybook, UnoCSS, and ESLint among others. Each upgrade has been tested to ensure that there are no breaking changes affecting the current functionalities. This step is crucial for maintaining the robustness of our environment and keeping up with recent advancements in the dependencies we rely on.