chore: update .gitignore and eslint config
Added `.eslintcache` to the `.gitignore` to prevent the ESLint cache file from being tracked by Git. This change ensures that the cache file, which is specific to individual development environments, does not clutter the repository.

Also updated the ESLint configuration to disable the 'sonarjs/sonar-no-unused-vars' rule. This adjustment aligns with our current coding standards and reduces noise in linting feedback for developers.